Rename shared pubspec testing file
