+++ test/dart2_suggestors/pubspec_over_react_upgrader_test.dart
@@ -16,7 +16,7 @@ import 'package:over_react_codemod/src/dart2_suggestors/pubspec_over_react_upgra
 import 'package:pub_semver/pub_semver.dart';
 import 'package:test/test.dart';
 
-import '../shared_pubspec_test.dart';
+import '../shared_pubspec_tests.dart';
 import '../util.dart';
 
 main() {
